refactor(layers) 🔨: Moved common files into a new dir
