+++ src/python/adapters/dynamodb_adapter.py
