+++ src/python/adapters/uefa_adapter.py
