+++ src/python/core/uefa_service.py
@@ -1,5 +1,5 @@
 import time
-from src.ports.uefa_port import UEFAPlayerStatsRepository
+from ports.uefa_port import UEFAPlayerStatsRepository
 from core.measurement_service import MeasurementService
 from data.ap2 import PlayerTotalScore
 from data.ap1 import PlayerMatchStats
